0024393: Visualization - objects position with enchanced precision
Graphic3d_ZLayerSettings - public fields have been replaced by methods.
Added new property Origin defining local coordinate system for all Layer objects.

Syntax of Draw Harness command VZLayer has been redesigned.

Graphic3d_CStructure now stores bounding box with double precision floats.
OpenGl_BVHTreeSelector - frustum culling now works with double precision floats.

+++ src/OpenGl/OpenGl_View_Raytrace.cxx
@@ -2239,9 +2239,9 @@ Standard_Boolean OpenGl_View::updateRaytraceLightSources (const OpenGl_Mat4& the
 
     if (aLight.Type != Graphic3d_TOLS_DIRECTIONAL)
     {
-      aPosition = BVH_Vec4f (aLight.Position.x(),
-                             aLight.Position.y(),
-                             aLight.Position.z(),
+      aPosition = BVH_Vec4f (static_cast<float>(aLight.Position.x()),
+                             static_cast<float>(aLight.Position.y()),
+                             static_cast<float>(aLight.Position.z()),
                              1.0f);
 
       // store smoothing radius in w-component
